Add files via upload
This system is a website for a company named “Poonam Water Suppliers” which basically deals with distribution of water tankers, water bottles and water cans. They do so for a single delivery as well as for repeating orders. This system aims to digitalize the workings of the company, attracting new customers and making the work process much easier than before. It also makes record keeping simple and efficient.